+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ## [Unreleased]
 
 ### Features
+* (build) [\#126](https://github.com/line/lbm/pull/126) Automatically generates release note and binaries
 
 ### Improvements